+++ src-tauri/src/protocols/mod.rs
@@ -1,3 +1,4 @@
 pub mod dash;
 pub mod hls;
+mod hls_options;
 pub mod strategy;